Make digest headers links
Needed some crazy workarounds as I was treating tags as functions. They do have
`apply' methods so this is fine, until I tried to interpret headers using an
actual function. It now works, I changed it so it always uses the apply method
as a function. No idea if what I've done is best or even really makes sense.